+++ CHANGELOG.md
@@ -3,8 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
-## [1.7.17] - unreleased
+## [1.8.0] - unreleased
 ### Added
+- Exit 1 when any issue messages are produced; thanks @i3v
+- GHORG_EXIT_CODE_ON_CLONE_INFOS to allow for control of exit code when any info messages are produced
 ### Changed
 ### Deprecated
 ### Removed